% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecreuset_palette.R
\docType{data}
\name{lc_palettes}
\alias{lc_palettes}
\title{List of color palettes.}
\format{
An object of class \code{list} of length 1.
}
\usage{
lc_palettes
}
\description{
Use \code{\link{render_pitt_palette}} to construct palettes of desired length.
}
\keyword{datasets}
