% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecreuset_palette.R
\name{render_lc_palette}
\alias{render_lc_palette}
\title{Le Creuset palette rendering}
\usage{
render_lc_palette(name, n)
}
\arguments{
\item{name}{Name of color palette. The only option is: \code{"colors_2024"}.}

\item{n}{Number of colors desired. Most palettes have 3-6 colors. Most color
schemes are derived from \href{https://www.brand.pitt.edu/visual-identity/color-palettes-0}.
If omitted, uses all colors.}
}
\value{
A vector of colors.
}
\description{
Le Creuset palette rendering
}
\examples{
render_lc_palette("colors_2024")
}
\keyword{colors}
